requirements.txt
celery
django-environ
djangorestframework
drf-spectacular
psycopg2-binary
redis
